+++ Kal/act3/src/com/kal/TwoDayPackage.java
@@ -5,17 +5,17 @@ public class TwoDayPackage extends Package{
     private int senderZip;
     private String recipientName, recipientAddress, recipientCity, recipientState;
     private int recipientZip;
-    private static float packWeight;
-    private static float packCost;
-    float totalCost;
-    private static float flatFee;
-    public TwoDayPackage(String senderName, String senderAddress, String senderCity, String senderState, int senderZip, String recipientName, String recipientAddress, String recipientCity, String recipientState, int recipientZip, float packageWeight, float packageCost, float flatFee) {
+    private static double packWeight;
+    private static double packCost;
+    double totalCost;
+    private static double flatFee;
+    public TwoDayPackage(String senderName, String senderAddress, String senderCity, String senderState, int senderZip, String recipientName, String recipientAddress, String recipientCity, String recipientState, int recipientZip, double packageWeight, double packageCost, double flatFee) {
         super(senderName, senderAddress, senderCity, senderState, senderZip, recipientName, recipientAddress, recipientCity, recipientState, recipientZip, packageWeight, packageCost);
         this.senderName = senderName; this.senderAddress = senderAddress; this.senderCity = senderCity; this.senderState = senderState; this.senderZip = senderZip;
         this.recipientName = recipientName; this.recipientAddress = recipientAddress; this.recipientCity = recipientCity; this.recipientState = recipientState; this.recipientZip = recipientZip;
         packCost = packageCost; packWeight = packageWeight; TwoDayPackage.flatFee = flatFee; totalCost = totalCalcCost();
     }
-    public static float totalCalcCost(){
+    public static double totalCalcCost(){
         return packCost*packWeight+flatFee;
     }
